+++ test/integration/braintree/dispute_search_spec.js
@@ -159,6 +159,25 @@ describe('DisputeSearch', () => {
       });
     });
 
+    it('returns disputes by chargebackProtectionLevel', (done) => {
+          let stream = specHelper.defaultGateway.dispute.search(function (search) {
+            return search.chargebackProtectionLevel().in([
+              Dispute.ChargebackProtectionLevel.Effortless
+            ]);
+          });
+
+          stream.on('data', dispute => disputes.push(dispute));
+
+          stream.on('end', () => {
+            assert(disputes.length == 1);
+            assert(disputes[0].caseNumber === "CASE-CHARGEBACK-PROTECTED");
+            assert(disputes[0].reason === Dispute.Reason.Fraud);
+            assert(disputes[0].chargebackProtectionLevel === Dispute.ChargebackProtectionLevel.Effortless);
+
+            done();
+          });
+        });
+
     it('returns disputes by customer_id', (done) => {
       specHelper.defaultGateway.customer.create({
         firstName: 'Action',